add fleshed out file contract test
