+++ modules/consensus/accept_test.go
@@ -1,7 +1,6 @@
 package consensus
 
 import (
-	"bytes"
 	"crypto/rand"
 	"errors"
 	"fmt"
@@ -370,255 +369,6 @@ func TestInconsistentCheck(t *testing.T) {
 	cst.miner.AddBlock()
 }
 */
-
-// testFileContractsBlocks creates a series of blocks that create, revise,
-// prove, and fail to prove file contracts.
-func (cst *consensusSetTester) testFileContractsBlocks() error {
-	var validProofDest, missedProofDest, revisionDest types.UnlockHash
-	_, err := rand.Read(validProofDest[:])
-	if err != nil {
-		return err
-	}
-	_, err = rand.Read(missedProofDest[:])
-	if err != nil {
-		return err
-	}
-	_, err = rand.Read(revisionDest[:])
-	if err != nil {
-		return err
-	}
-
-	// Create a file (as a bytes.Buffer) that will be used for file contracts
-	// and storage proofs.
-	filesize := uint64(4e3)
-	fileBytes := make([]byte, filesize)
-	_, err = rand.Read(fileBytes)
-	if err != nil {
-		return err
-	}
-	file := bytes.NewReader(fileBytes)
-	merkleRoot, err := crypto.ReaderMerkleRoot(file)
-	if err != nil {
-		return err
-	}
-	file.Seek(0, 0)
-
-	// Create a file contract that will be successfully proven and an alternate
-	// file contract which will be missed.
-	payout := types.NewCurrency64(400e6)
-	validFC := types.FileContract{
-		FileSize:       filesize,
-		FileMerkleRoot: merkleRoot,
-		WindowStart:    cst.cs.height() + 2,
-		WindowEnd:      cst.cs.height() + 4,
-		Payout:         payout,
-		ValidProofOutputs: []types.SiacoinOutput{{
-			UnlockHash: validProofDest,
-		}},
-		MissedProofOutputs: []types.SiacoinOutput{{
-			UnlockHash: missedProofDest,
-		}},
-		UnlockHash: types.UnlockConditions{}.UnlockHash(),
-	}
-	outputSize := payout.Sub(types.Tax(cst.cs.dbBlockHeight(), validFC.Payout))
-	validFC.ValidProofOutputs[0].Value = outputSize
-	validFC.MissedProofOutputs[0].Value = outputSize
-	missedFC := types.FileContract{
-		FileSize:       uint64(filesize),
-		FileMerkleRoot: merkleRoot,
-		WindowStart:    cst.cs.height() + 2,
-		WindowEnd:      cst.cs.height() + 4,
-		Payout:         payout,
-		ValidProofOutputs: []types.SiacoinOutput{{
-			Value:      outputSize,
-			UnlockHash: validProofDest,
-		}},
-		MissedProofOutputs: []types.SiacoinOutput{{
-			Value:      outputSize,
-			UnlockHash: missedProofDest,
-		}},
-		UnlockHash: types.UnlockConditions{}.UnlockHash(),
-	}
-
-	// Create and fund a transaction with a file contract.
-	txnBuilder := cst.wallet.StartTransaction()
-	err = txnBuilder.FundSiacoins(payout.Mul(types.NewCurrency64(2)))
-	if err != nil {
-		return err
-	}
-	validFCIndex := txnBuilder.AddFileContract(validFC)
-	missedFCIndex := txnBuilder.AddFileContract(missedFC)
-	txnSet, err := txnBuilder.Sign(true)
-	if err != nil {
-		return err
-	}
-	ti := len(txnSet) - 1
-	validFCID := txnSet[ti].FileContractID(int(validFCIndex))
-	missedFCID := txnSet[ti].FileContractID(int(missedFCIndex))
-	err = cst.tpool.AcceptTransactionSet(txnSet)
-	if err != nil {
-		return err
-	}
-	_, err = cst.miner.AddBlock()
-	if err != nil {
-		return err
-	}
-
-	// Check that the siafund pool was increased.
-	var siafundPool types.Currency
-	err = cst.cs.db.Update(func(tx *bolt.Tx) error {
-		siafundPool = getSiafundPool(tx)
-		return nil
-	})
-	if err != nil {
-		panic(err)
-	}
-	if siafundPool.Cmp(types.NewCurrency64(31200e3)) != 0 {
-		return errors.New("siafund pool was not increased correctly")
-	}
-
-	// Submit a file contract revision to the missed-proof file contract.
-	txn := types.Transaction{
-		FileContractRevisions: []types.FileContractRevision{{
-			ParentID:          missedFCID,
-			NewRevisionNumber: 1,
-
-			NewFileSize:          10e3, // By changing the filesize without changing the hash, a proof should become impossible.
-			NewFileMerkleRoot:    missedFC.FileMerkleRoot,
-			NewWindowStart:       missedFC.WindowStart + 1,
-			NewWindowEnd:         missedFC.WindowEnd,
-			NewValidProofOutputs: missedFC.ValidProofOutputs,
-			NewMissedProofOutputs: []types.SiacoinOutput{{
-				Value:      outputSize,
-				UnlockHash: revisionDest,
-			}},
-		}},
-	}
-	err = cst.tpool.AcceptTransactionSet([]types.Transaction{txn})
-	if err != nil {
-		return err
-	}
-	_, err = cst.miner.AddBlock()
-	if err != nil {
-		return err
-	}
-
-	// Check that the revision was successful.
-	if cst.cs.db.getFileContracts(missedFCID).RevisionNumber != 1 {
-		return errors.New("revision did not update revision number")
-	}
-	if cst.cs.db.getFileContracts(missedFCID).FileSize != 10e3 {
-		return errors.New("revision did not update file contract size")
-	}
-
-	// Create a storage proof for the validFC and submit it in a block.
-	spSegmentIndex, err := cst.cs.StorageProofSegment(validFCID)
-	if err != nil {
-		return err
-	}
-	segment, hashSet, err := crypto.BuildReaderProof(file, spSegmentIndex)
-	if err != nil {
-		return err
-	}
-	txn = types.Transaction{
-		StorageProofs: []types.StorageProof{{
-			ParentID: validFCID,
-			HashSet:  hashSet,
-		}},
-	}
-	copy(txn.StorageProofs[0].Segment[:], segment)
-	err = cst.tpool.AcceptTransactionSet([]types.Transaction{txn})
-	if err != nil {
-		return err
-	}
-	_, err = cst.miner.AddBlock()
-	lockID3 := cst.cs.mu.Lock()
-	cst.cs.mu.Unlock(lockID3)
-
-	if err != nil {
-		return err
-	}
-
-	// Check that the valid contract was removed but the missed contract was
-	// not.
-	exists := cst.cs.db.inFileContracts(validFCID)
-	if exists {
-		return errors.New("valid file contract still exists in the consensus set")
-	}
-	exists = cst.cs.db.inFileContracts(missedFCID)
-	if !exists {
-		return errors.New("missed file contract was consumed by storage proof")
-	}
-
-	// Check that the file contract output made it into the set of delayed
-	// outputs.
-	validProofID := validFCID.StorageProofOutputID(types.ProofValid, 0)
-	exists = cst.cs.db.inDelayedSiacoinOutputsHeight(cst.cs.height()+types.MaturityDelay, validProofID)
-	if !exists {
-		return errors.New("file contract payout is not in the delayed outputs set")
-	}
-
-	// Mine a block to close the window on the missed file contract.
-	_, err = cst.miner.AddBlock()
-	if err != nil {
-		return err
-	}
-	exists = cst.cs.db.inFileContracts(validFCID)
-	if exists {
-		return errors.New("valid file contract still exists in the consensus set")
-	}
-	exists = cst.cs.db.inFileContracts(missedFCID)
-	if exists {
-		return errors.New("missed file contract was not consumed when the window was closed.")
-	}
-
-	// Mine enough blocks to get all of the outputs into the set of siacoin
-	// outputs.
-	for i := types.BlockHeight(0); i <= types.MaturityDelay; i++ {
-		_, err = cst.miner.AddBlock()
-		if err != nil {
-			return err
-		}
-	}
-
-	// Check that all of the outputs have ended up at the right destination.
-	if cst.cs.db.getSiacoinOutputs(validFCID.StorageProofOutputID(types.ProofValid, 0)).UnlockHash != validProofDest {
-		return errors.New("file contract output did not end up at the right place.")
-	}
-	if cst.cs.db.getSiacoinOutputs(missedFCID.StorageProofOutputID(types.ProofMissed, 0)).UnlockHash != revisionDest {
-		return errors.New("missed file proof output did not end up at the revised destination")
-	}
-
-	return nil
-}
-
-// TestFileContractsBlocks creates a consensus set tester and uses it to call
-// testFileContractsBlocks.
-func TestFileContractsBlocks(t *testing.T) {
-	if testing.Short() {
-		t.SkipNow()
-	}
-	cst, err := createConsensusSetTester("TestFileContractsBlocks")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer cst.closeCst()
-
-	// COMPATv0.4.0
-	//
-	// Mine enough blocks to get above the file contract hardfork threshold
-	// (10).
-	for i := 0; i < 10; i++ {
-		_, err = cst.miner.AddBlock()
-		if err != nil {
-			t.Fatal(err)
-		}
-	}
-	err = cst.testFileContractsBlocks()
-	if err != nil {
-		t.Fatal(err)
-	}
-}
 
 // testSpendSiafundsBlock mines a block with a transaction spending siafunds
 // and adds it to the consensus set.
