+++ modules/consensus/consensusset_test.go
@@ -1,6 +1,7 @@
 package consensus
 
 import (
+	"bytes"
 	"crypto/rand"
 	"path/filepath"
 	"testing"
@@ -38,6 +39,17 @@ func randAddress() types.UnlockHash {
 		panic(err)
 	}
 	return uh
+}
+
+// randFile returns a bytes.Reader that is equivalent to a random file of size
+// 'filesize'.
+func randFile(filesize uint64) *bytes.Reader {
+	fileBytes := make([]byte, filesize)
+	_, err := rand.Read(fileBytes)
+	if err != nil {
+		panic(err)
+	}
+	return bytes.NewReader(fileBytes)
 }
 
 // createConsensusSetTester creates a consensusSetTester that's ready for use.
